+++ src/core/highlighter.rs
@@ -88,7 +88,6 @@ impl Default for Highlighter {
             .with_date_time_highlighters(DateTimeConfig::default())
             .with_url_highlighter(UrlConfig::default())
             .with_ip_v4_highlighter(IpV4Config::default())
-            .with_ip_v6_highlighter(IpV6Config::default())
             .with_uuid_highlighter(UuidConfig::default())
             .with_pointer_highlighter(PointerConfig::default())
             .with_unix_path_highlighter(UnixPathConfig::default())